add sweetalert2

+++ resources/js/components/PostComponent.vue
@@ -2,10 +2,11 @@
     <div class="container">
         <div class="row">
             <div class="col-md-12">
-                        <p>My Post</p>
-                        <button @click="initAddPost()" class="btn btn-primary btn-xs pull-right"> + Add New Post </button>
+                    <div class="mt-5">
+                        <button @click="initAddPost()" class="btn btn-primary btn-xs"> + Add New Post </button>
+                    </div>
 
-                        <table class="table table-bordered table-striped table-responsive" v-if="posts.length > 0">
+                        <table class="table table-bordered table-striped table-responsive mt-5" v-if="posts.length > 0">
                             <tbody>
                                 <tr>
                                     <th>
@@ -114,9 +115,12 @@
 </template>
 
 <script>
+import Swal from 'sweetalert2';
     export default {
         data(){
             return {
+                status: '',
+                message: '',
                 posts: {
                     title: '',
                     description: ''
@@ -131,6 +135,13 @@
             this.readPosts();
         },
         methods: {
+            showAlert(status, message){
+                this.$swal(
+                    'Exito!',
+                    message,
+                    status
+                );
+            },
             initAddPost()
             {
                 this.errors = [];
@@ -148,6 +159,7 @@
 
                         $("#add_post_model").modal("hide");
                         this.readPosts();
+                        this.showAlert('success','Registro Exitoso');
 
                     })
                     .catch(error => {
@@ -205,26 +217,28 @@
             },
             deletePost(index)
             {
-                //let conf = confirm("Do you ready want to delete this post?");
-                //if (conf === true) {
-
-                    axios.delete('/posts/' + this.posts[index].id)
-                        .then(response => {
-
+                this.$swal({
+                title: 'Esta seguro de borrarlo?',
+                text: "No lo podra recuperar",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Si, Borralo!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        axios.delete('/posts/' + this.posts[index].id)
+                            .then(response => {
                             this.posts.splice(index, 1);
-
-                        })
-                        .catch(error => {
-                            this.errors = [];
-                            if (error.response.data.errors.title) {
-                                this.errors.push(error.response.data.errors.title[0]);
-                            }
-
-                            if (error.response.data.errors.description) {
-                                this.errors.push(error.response.data.errors.description[0]);
-                            }
                         });
-                //}
+                        Swal.fire(
+                        'Borrado!',
+                        'El registro se elimino exitosamente.',
+                        'success'
+                        );
+
+                    }
+                });
             }
         }
     }
